refactor file tree
