+++ src/App.js
@@ -1,0 +1,12 @@
+// import './App.css';
+import NavBar from './components/NavBar/navBar';
+
+function App() {
+  return (
+    <div className="App">
+      <NavBar></NavBar>
+    </div>
+  );
+}
+
+export default App;
