--- src/Components/App/App.js
@@ -1,30 +1,0 @@
-import logo from '../../logo.svg';
-import './App.css';
-
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <p>hi</p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-       
-          
-          <h1 className="text-3xl font-bold underline">
-               Working Tailwind!
-           </h1>
-        </a>
-      </header>
-    </div>
-  );
-}
-
-export default App;
